Create separate log formatter config file
