+++ docker/Dockerfile-ci
@@ -130,7 +130,7 @@ RUN \
   /bin/sh -c 'cd packages/backend && EXIT_AFTER_READY=1 PORT_OFFSET=4096 node -r ./test/env.ts dist/server.mjs' && \
   # Create directories for target container and copy only necessary files
   mkdir -p /app/public/api-docs /app/prisma/client && \
-  cp packages/backend/dist/server.mjs* /app && \
+  cp packages/backend/dist/* /app && \
   cp -r packages/backend/prisma/client/schema.prisma packages/backend/prisma/client/libquery_engine-* /app/prisma/client && \
   cp -r packages/frontend/dist/* /app/public && \
   # Copy project files as well as Swagger UI files and WebAssembly package
