+++ packages/backend/package.json
@@ -25,7 +25,8 @@
     "prisma:studio": "pnpify prisma studio",
     "prisma:generate": "pnpify prisma generate",
     "prisma:migrate:dev": "pnpify prisma migrate dev",
-    "build": "esbuild --color=true src/server.ts --bundle --platform=node --format=esm --legal-comments=none --sourcemap=linked --sources-content=false --banner:js=\"import __path from'path';import{createRequire as __createRequire}from'module';import{fileURLToPath}from'url';const require=__createRequire(import.meta.url),__filename=fileURLToPath(import.meta.url),__dirname=__path.dirname(__filename);\" --outfile=dist/server.mjs",
+    "build": "esbuild --color=true src/server.ts --bundle --platform=node --format=esm --legal-comments=none --sourcemap=linked --sources-content=false --banner:js=\"import __path from'path';import{createRequire as __createRequire}from'module';import{fileURLToPath}from'url';const require=__createRequire(import.meta.url),__filename=fileURLToPath(import.meta.url),__dirname=__path.dirname(__filename);\" --log-override:commonjs-variable-in-esm=silent --outfile=dist/server.mjs && yarn build:logFormatterConfig",
+    "build:logFormatterConfig": "esbuild src/utils/logFormatterConfig.ts --bundle --minify --platform=node --format=cjs --legal-comments=none --log-override:commonjs-variable-in-esm=silent --outfile=dist/pino-pretty-config.cjs",
     "typecheck": "tsc --noEmit --tsBuildInfoFile $HOME/.cache/rating-tracker/backend.tsbuildinfo",
     "lint": "eslint --color --cache --cache-location $HOME/.cache/rating-tracker/backend.eslintcache --cache-strategy content --max-warnings 0 --ext .ts src/ test/",
     "lint:fix": "yarn lint --fix"
